+++ macos/Classes/AudioPlayer.h
@@ -1,0 +1,16 @@
+#import <FlutterMacOS/FlutterMacOS.h>
+
+@interface AudioPlayer : NSObject<FlutterStreamHandler>
+
+- (instancetype)initWithRegistrar:(NSObject<FlutterPluginRegistrar> *)registrar playerId:(NSString*)idParam;
+
+@end
+
+enum PlaybackState {
+	none,
+	stopped,
+	paused,
+	playing,
+	connecting,
+	completed
+};